chore(graphics): file split
